Modle testing complete

+++ Naive Bayes/Naive Bayes/Model.hpp
@@ -25,7 +25,7 @@ private:
     friend std::istream& operator >> (std::istream& is, Model& model);
 public:
     bool LoadModel(std::string file_location);
-    bool Train(ImageSet set, int smooth_factor);
+    bool Train(ImageSet set, double smooth_factor);
     int Evaluate(Image image);
 };
